backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
rembg>=2.0.50
Pillow>=10.0.0